renderer/pyproject.toml
[tool.poetry]
name = "renderer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.112.2"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
docker = "^7.1.0"
python-dotenv = "^1.0.1"
oci = "^2.133.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
